# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Live Chat API
  version: 1.0.0
  description: >
    Documentation Live Chat. 
    L'API expose l'authentification JWT, la gestion des profils, des serveurs,
    des salons, des membres, des conversations privees et des messages texte.
    Les messages sont stockes dans MongoDB, tandis que les donnees relationnelles
    passent par Prisma/PostgreSQL.
servers:
  - url: http://localhost:3000
    description: Backend local
...
security:
  - bearerAuth: []
paths:
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Obtenir le profil courant.
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profil courant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
              examples:
                success:
                  value:
                    user:
                      id: 12
                      pseudo: alex
                      email: alex@example.com
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      tags: [Profile]
      summary: Mettre a jour le profil courant.
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
            examples:
              changePseudo:
                value:
                  pseudo: alex_dev
              changePassword:
                value:
                  password: NewStrongPassword123!
                  currentPassword: StrongPassword123!
      responses:
        '200':
          description: Profil mis a jour.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Profile]
      summary: Supprimer le compte courant.
      operationId: deleteProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileDeleteRequest'
            examples:
              example:
                value:
                  password: StrongPassword123!
      responses:
        '204':
          description: Compte supprime.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/servers:
    get:
      tags: [Servers]
      summary: Lister les serveurs accessibles au compte courant.
      operationId: listServers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des serveurs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServersListResponse'
              examples:
                success:
                  value:
                    servers:
                      - id: 1
                        name: General
                        icon_url: null
                        owner_id: 12
                        created_at: '2026-08-01T10:00:00.000Z'
                        channels:
                          - id: 5
                            name: général
                            type: text
                            server_id: 1
                            created_at: '2026-08-01T10:00:01.000Z'
                        members:
                          - user_id: 12
                            role: admin
                        _count:
                          members: 4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Servers]
      summary: Creer un serveur.
      operationId: createServer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreateRequest'
            examples:
              example:
                value:
                  name: Projet Produit
                  icon_url: https://cdn.example.com/icons/server.png
      responses:
        '201':
          description: Serveur cree.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/servers/{serverId}/channels:
    post:
      tags: [Servers]
      summary: Ajouter un salon a un serveur.
      operationId: createServerChannel
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelCreateRequest'
            examples:
              example:
                value:
                  name: support
                  type: text
      responses:
        '201':
          description: Salon cree.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/servers/{serverId}/members:
    post:
      tags: [Members]
      summary: Ajouter un membre a un serveur.
      operationId: addServerMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerMemberCreateRequest'
            examples:
              example:
                value:
                  pseudo: marina
      responses:
        '201':
          description: Membre ajoute.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerMemberCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags: [Members]
      summary: Lister tous les membres d'un serveur.
      operationId: listServerMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServerId'
      responses:
        '200':
          description: Liste des membres.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerMembersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/conversations:
    get:
      tags: [Conversations]
      summary: Lister les conversations privees du compte courant.
      operationId: listConversations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conversations du compte.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Conversations]
      summary: Creer ou ouvrir une conversation privee.
      operationId: createConversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreateRequest'
            examples:
              example:
                value:
                  pseudo: marina
      responses:
        '201':
          description: Conversation creee ou recuperee.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/users/search:
    get:
      tags: [Users]
      summary: Rechercher des utilisateurs par pseudo.
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Terme de recherche applique au pseudo.
          schema:
            type: string
      responses:
        '200':
          description: Liste des utilisateurs correspondants.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/users/status:
    get:
      tags: [Users]
      summary: Lire le statut courant du compte.
      operationId: getCurrentUserStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statut manuel et statut effectif.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ServerId:
      name: serverId
      in: path
      required: true
      description: Identifiant numerique du serveur.
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Requete invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            example:
              value:
                message: Parametres invalides.
    Unauthorized:
      description: Authentification manquante ou invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            example:
              value:
                message: Token requis.
    Forbidden:
      description: Authentifie mais non autorise.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            example:
              value:
                message: Accès refusé.
    NotFound:
      description: Ressource introuvable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            example:
              value:
                message: Ressource introuvable.
    Conflict:
      description: Conflit sur la ressource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            example:
              value:
                message: Cet utilisateur est déjà membre du serveur.
    ServerError:
      description: Erreur serveur inattendue.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            example:
              value:
                message: Erreur serveur.
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ProfileResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: integer
            pseudo:
              type: string
            email:
              type: string
              format: email
          required: [id, pseudo, email]
      required: [user]
    ProfileUpdateRequest:
      type: object
      properties:
        pseudo:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        currentPassword:
          type: string
          format: password
      additionalProperties: false
    ProfileDeleteRequest:
      type: object
      properties:
        password:
          type: string
          format: password
      required: [password]
...
    ChannelType:
      type: string
      enum: [text, voice]
    ServerMemberRole:
      type: string
      enum: [admin, member]
    StatusValue:
      type: string
      enum: [online, away, dnd, offline]
    Server:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: General
        icon_url:
          type: string
          nullable: true
          example: null
        owner_id:
          type: integer
          example: 12
        created_at:
          type: string
          format: date-time
      required: [id, name, icon_url, owner_id, created_at]
    Channel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          $ref: '#/components/schemas/ChannelType'
        server_id:
          type: integer
        created_at:
          type: string
          format: date-time
      required: [id, name, type, server_id, created_at]
    ServerMember:
      type: object
      properties:
        user_id:
          type: integer
        server_id:
          type: integer
        role:
          $ref: '#/components/schemas/ServerMemberRole'
        joined_at:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: integer
            pseudo:
              type: string
            email:
              type: string
              format: email
          required: [id, pseudo, email]
      required: [user_id, server_id, role, joined_at]
    ServerListItem:
      allOf:
        - $ref: '#/components/schemas/Server'
        - type: object
          properties:
            channels:
              type: array
              items:
                $ref: '#/components/schemas/Channel'
            members:
              type: array
              items:
                type: object
                properties:
                  user_id:
                    type: integer
                  role:
                    $ref: '#/components/schemas/ServerMemberRole'
                required: [user_id, role]
            _count:
              type: object
              properties:
                members:
                  type: integer
              required: [members]
    ServersListResponse:
      type: object
      properties:
        servers:
          type: array
          items:
            $ref: '#/components/schemas/ServerListItem'
      required: [servers]
    ServerCreateRequest:
      type: object
      properties:
        name:
          type: string
          example: Projet Produit
        icon_url:
          type: string
          nullable: true
          example: https://cdn.example.com/icons/server.png
      required: [name]
    ServerCreateResponse:
      type: object
      properties:
        server:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            icon_url:
              type: string
              nullable: true
            owner_id:
              type: integer
            created_at:
              type: string
              format: date-time
            channels:
              type: array
              items:
                $ref: '#/components/schemas/Channel'
          required: [id, name, icon_url, owner_id, created_at, channels]
      required: [server]
    ChannelCreateRequest:
      type: object
      properties:
        name:
          type: string
          example: support
        type:
          $ref: '#/components/schemas/ChannelType'
      required: [name]
    ChannelCreateResponse:
      type: object
      properties:
        channel:
          $ref: '#/components/schemas/Channel'
      required: [channel]
...
    ServerMemberCreateRequest:
      type: object
      properties:
        pseudo:
          type: string
          example: marina
      required: [pseudo]
    ServerMemberCreateResponse:
      type: object
      properties:
        member:
          $ref: '#/components/schemas/ServerMember'
      required: [member]
    ServerMembersResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/ServerMember'
      required: [members]
...
    ConversationUserView:
      type: object
      properties:
        id:
          type: integer
        pseudo:
          type: string
        email:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/StatusValue'
      required: [id, pseudo, email, status]
    ConversationListItem:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/ConversationUserView'
      required: [id, createdAt, user]
    ConversationsListResponse:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/ConversationListItem'
      required: [conversations]
    ConversationCreateRequest:
      type: object
      properties:
        pseudo:
          type: string
          example: marina
      required: [pseudo]
    ConversationCreateResponse:
      type: object
      properties:
        conversation:
          $ref: '#/components/schemas/ConversationListItem'
      required: [conversation]
...
    UserSearchResult:
      type: object
      properties:
        id:
          type: integer
        pseudo:
          type: string
        status:
          $ref: '#/components/schemas/StatusValue'
      required: [id, pseudo, status]
    UserSearchResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserSearchResult'
      required: [users]
    UserStatusResponse:
      type: object
      properties:
        manualStatus:
          $ref: '#/components/schemas/StatusValue'
        status:
          $ref: '#/components/schemas/StatusValue'
      required: [manualStatus, status]
...